filmpalast.to/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyYAML>=6.0
fake-useragent>=1.4.0
pandas>=2.0.0
openpyxl>=3.1.0
